requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
Django==4.0.1
djangorestframework==3.13.1
idna==3.3
pycparser==2.21
PyJWT==2.3.0
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.8
